# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tunnel Management API",
    "description": "API for managing tunnels",
    "version": "1.0."
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "API key is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "state.auth.fail"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
    "/": {
      "get": {
        "security": [],
        "summary": "Get info about the api",
        "responses": {
          "200": {
            "description": "Info",
            "content": {
              "application/json": {
                "examples": {
                  "Tunnel Running": {
                    "value": {
                      "docs": "/docs/",
                      "status": "state.tunnel.running"
                    }
                  },
                  "Tunnel Stopped": {
                    "value": {
                      "docs": "/docs/",
                      "status": "state.tunnel.stopped"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tunnels/list": {
      "get": {
        "security": [
          {
            "apiKey": []
          }
        ],
        "summary": "List tunnels",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "Test"
                      },
                      "host": {
                        "type": "string",
                        "example": "127.0.0.1:9000"
                      },
                      "target": {
                        "type": "string",
                        "example": "192.168.0.2:9000"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/tunnels/restart": {
      "post": {
        "summary": "Restart tunnels",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.restarted"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/tunnels/start": {
      "post": {
        "summary": "Start tunnels",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.started"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/tunnels/stop": {
      "post": {
        "summary": "Stop tunnels",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.stopped"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/tunnels/delete/{id}": {
      "post": {
        "summary": "Delete tunnel by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "test"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tunnel deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.tunnel.delete.deleted"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Tunnel not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.tunnel.delete.notfound"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/tunnels/edit/{id}": {
      "post": {
        "summary": "Edit tunnel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "test"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Test",
                  "target": "127.0.0.1:8888",
                  "host": "192.168.0.1:8888"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tunnel edited",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.tunnel.edit.success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Edit failed (Invalid data)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.tunnel.edit.invalid.data"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Edit failed (Tunnel Not Found)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.tunnel.edit.invalid.notfound"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/tunnels/create/{id}": {
      "post": {
        "summary": "Create tunnel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "test"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Test"
                  },
                  "host": {
                    "type": "string",
                    "example": "127.0.0.1:9000"
                  },
                  "target": {
                    "type": "string",
                    "example": "192.168.0.2:9000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tunnel created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.tunnel.create.success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid data (missing required fields)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.tunnel.create.invalid.data"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid data (duplicate tunnel name)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "state.tunnel.create.invalid.duplicate"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    }
  }
}